LPS-65111 Create a new UpgradeModules implementation to rerun the the logic in 7.0.1
