+++ portal-impl/src/com/liferay/portal/upgrade/UpgradeProcess_7_0_1.java
@@ -18,6 +18,7 @@ import com.liferay.portal.kernel.upgrade.UpgradeProcess;
 import com.liferay.portal.kernel.util.ReleaseInfo;
 import com.liferay.portal.upgrade.v7_0_1.UpgradeCompany;
 import com.liferay.portal.upgrade.v7_0_1.UpgradeMessageBoards;
+import com.liferay.portal.upgrade.v7_0_1.UpgradeModules;
 
 /**
  * @author Brian Wing Shun Chan
@@ -33,6 +34,7 @@ public class UpgradeProcess_7_0_1 extends UpgradeProcess {
 	protected void doUpgrade() throws Exception {
 		upgrade(UpgradeCompany.class);
 		upgrade(UpgradeMessageBoards.class);
+		upgrade(UpgradeModules.class);
 
 		clearIndexesCache();
 	}
