+++ portal-impl/src/com/liferay/portal/upgrade/v7_0_1/UpgradeModules.java
@@ -1,0 +1,36 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.upgrade.v7_0_1;
+
+/**
+ * @author Roberto Díaz
+ */
+public class UpgradeModules
+	extends com.liferay.portal.upgrade.util.UpgradeModules {
+
+	@Override
+	public String[] getBundleSymbolicNames() {
+		return _bundleSymbolicNames;
+	}
+
+	@Override
+	public String[][] getConvertedLegacyModules() {
+		return _convertedLegacyModules;
+	}
+
+	private static final String[] _bundleSymbolicNames = new String[0];
+	private static final String[][] _convertedLegacyModules = {};
+
+}
